+++ tomcat/src/main/java/nextstep/jwp/handler/HttpRequestMapping.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package nextstep.jwp.handler;
 
 import java.util.Objects;
 import org.apache.coyote.http11.HttpRequest;
