+++ tomcat/src/main/java/nextstep/jwp/handler/RequestHandlerMapping.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package nextstep.jwp.handler;
 
 import java.util.HashMap;
 import java.util.Map;
