+++ tomcat/src/main/java/org/apache/coyote/http11/SessionManager.java
@@ -1,6 +1,5 @@
-package org.apache.catalina;
+package org.apache.coyote.http11;
 
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
